Add files via upload
This project is a word guessing game, where the player tries to guess a randomly selected animal word by writing one letter at a time. The player has limited number of incorrect guesses, and for every incorrect guess a part of a  star is drawn, and if the stick figure is completed before the word is guessed correctly, the player loses the game.