Add files via upload
Nowadays, we are having much trouble with managing passwords and very often, we use same passwords for every website and thats why i made this. password creater and storer will create a random password for our username and website name we will enter and store it to a file in an encrypted manner (i have designed that encryption too instead of using a module) .& we will simply use that password on that website or app and now to extract that password when we want to use it next time from the file, we will mention our details in password extractor and it will decode the password and tell us the password.