Add files via upload
This is the simplest code ever. but yeah it has added a value to my life for sure. These are the things i make as snack in the evening and sometimes it gets hard to decide what to eat and i just run the program and eat a random thing or i saw the list and something popup that what i want to eat.